- WIP: added console script which uses overpass service. it should load/update the stored data about amenities.

+++ src/main/java/com/staysafe/dto/OverpassPoiDTO.java
@@ -1,6 +1,6 @@
 package com.staysafe.dto;
 
-public record OverpassPoiDTO(double id, String overpassType, double lat, double lon, String name) {
+public record OverpassPoiDTO(long id, String overpassType, double lat, double lon, String name) {
     @Override
     public String name() {
         return (name != null) ? name : "";
